requirements.txt
flask
google-generativeai
gunicorn
firebase-admin